Rename imports in components
