+++ assets/js/Components/Admin/AdminHeader.js
@@ -1,7 +1,7 @@
 import React from 'react'
 //import Logo from '../../../mediaAssets/logo_salmon.png'
 import Logo from '../../../mediaAssets/udoit_logo.png'
-import Classes from '../../../css/header.scss'
+import Classes from '../../../css/header.css'
 import { View } from '@instructure/ui-view'
 import { Text } from '@instructure/ui-text'
 import { AppNav } from '@instructure/ui-navigation'
