+++ assets/js/Components/Header.js
@@ -1,7 +1,7 @@
 import React from 'react'
 //import Logo from '../../mediaAssets/logo_salmon.png'
 import Logo from '../../mediaAssets/udoit_logo.png'
-import Classes from '../../css/header.scss'
+import Classes from '../../css/header.css'
 import { View } from '@instructure/ui-view'
 import { Flex } from '@instructure/ui-flex'
 import { IconButton } from '@instructure/ui-buttons'
